extracted factory functions
